double -> float, 3x speedup for cpp parallel A-KSVD

+++ Implementations/AK-SVD-Parallel/distorter.hpp
@@ -4,16 +4,16 @@
 class Distorter {
 public:
     // Adds Gaussian noise (scale noise_level) to the right half of image.
-    Distorter(const Eigen::MatrixXd& image, double noise_level = 0.075, unsigned int seed = 42);
+    Distorter(const Eigen::MatrixXf& image, float noise_level = 0.075, unsigned int seed = 42);
 
     // Extracts all overlapping patch_size×patch_size patches from the left half of the
     // distorted image, flattened row-major, then column-normalised (zero mean, unit std).
-    Eigen::MatrixXd extractNormalizedPatches(int patch_size) const;
+    Eigen::MatrixXf extractNormalizedPatches(int patch_size) const;
 
-    const Eigen::MatrixXd& getOriginal()  const { return original_;  }
-    const Eigen::MatrixXd& getDistorted() const { return distorted_; }
+    const Eigen::MatrixXf& getOriginal()  const { return original_;  }
+    const Eigen::MatrixXf& getDistorted() const { return distorted_; }
 
 private:
-    Eigen::MatrixXd original_;
-    Eigen::MatrixXd distorted_;
+    Eigen::MatrixXf original_;
+    Eigen::MatrixXf distorted_;
 };
