+++ Implementations/AK-SVD-Parallel/omp.hpp
@@ -16,13 +16,13 @@ class OMP {
         int N          = 0;
         int batch_size = 256;
     };
-    OMP(int N, int K, int T0, Eigen::MatrixXd Y, int batch_size = 256);
+    OMP(int N, int K, int T0, Eigen::MatrixXf Y, int batch_size = 256);
     Params params;
-    Eigen::MatrixXd Y;
+    Eigen::MatrixXf Y;
 
     bool do_omp();
 
-    Eigen::MatrixXd D;  // N x K: each column is a dictionary atom
-    Eigen::MatrixXd X;  // M x K: sparse coefficients (output)
+    Eigen::MatrixXf D;  // N x K: each column is a dictionary atom
+    Eigen::MatrixXf X;  // M x K: sparse coefficients (output)
 };
